# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8081","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/telemetry":{"get":{"tags":["telemetry-controller"],"operationId":"getTelemetrys","parameters":[{"name":"hostname","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TelemetryInfoDto"}}}}}}},"post":{"tags":["telemetry-controller"],"operationId":"postTelemetry","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelemetryDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","enum":["CONTINUE","SWITCHING_PROTOCOLS","PROCESSING","EARLY_HINTS","CHECKPOINT","OK","CREATED","ACCEPTED","NON_AUTHORITATIVE_INFORMATION","NO_CONTENT","RESET_CONTENT","PARTIAL_CONTENT","MULTI_STATUS","ALREADY_REPORTED","IM_USED","MULTIPLE_CHOICES","MOVED_PERMANENTLY","FOUND","MOVED_TEMPORARILY","SEE_OTHER","NOT_MODIFIED","USE_PROXY","TEMPORARY_REDIRECT","PERMANENT_REDIRECT","BAD_REQUEST","UNAUTHORIZED","PAYMENT_REQUIRED","FORBIDDEN","NOT_FOUND","METHOD_NOT_ALLOWED","NOT_ACCEPTABLE","PROXY_AUTHENTICATION_REQUIRED","REQUEST_TIMEOUT","CONFLICT","GONE","LENGTH_REQUIRED","PRECONDITION_FAILED","PAYLOAD_TOO_LARGE","REQUEST_ENTITY_TOO_LARGE","URI_TOO_LONG","REQUEST_URI_TOO_LONG","UNSUPPORTED_MEDIA_TYPE","REQUESTED_RANGE_NOT_SATISFIABLE","EXPECTATION_FAILED","I_AM_A_TEAPOT","INSUFFICIENT_SPACE_ON_RESOURCE","METHOD_FAILURE","DESTINATION_LOCKED","UNPROCESSABLE_ENTITY","LOCKED","FAILED_DEPENDENCY","TOO_EARLY","UPGRADE_REQUIRED","PRECONDITION_REQUIRED","TOO_MANY_REQUESTS","REQUEST_HEADER_FIELDS_TOO_LARGE","UNAVAILABLE_FOR_LEGAL_REASONS","INTERNAL_SERVER_ERROR","NOT_IMPLEMENTED","BAD_GATEWAY","SERVICE_UNAVAILABLE","GATEWAY_TIMEOUT","HTTP_VERSION_NOT_SUPPORTED","VARIANT_ALSO_NEGOTIATES","INSUFFICIENT_STORAGE","LOOP_DETECTED","BANDWIDTH_LIMIT_EXCEEDED","NOT_EXTENDED","NETWORK_AUTHENTICATION_REQUIRED"]}}}}}}},"/api/device":{"get":{"tags":["device-controller"],"operationId":"getAllByType","parameters":[{"name":"type","in":"query","required":false,"schema":{"type":"string","enum":["Laptop","Desktop","Tablet","Smatphone"]}},{"name":"lowThreshold","in":"query","required":false,"schema":{"type":"number","format":"double"}},{"name":"upThreshold","in":"query","required":false,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DeviceInfoDto"}}}}}}},"post":{"tags":["device-controller"],"operationId":"postDevice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","enum":["CONTINUE","SWITCHING_PROTOCOLS","PROCESSING","EARLY_HINTS","CHECKPOINT","OK","CREATED","ACCEPTED","NON_AUTHORITATIVE_INFORMATION","NO_CONTENT","RESET_CONTENT","PARTIAL_CONTENT","MULTI_STATUS","ALREADY_REPORTED","IM_USED","MULTIPLE_CHOICES","MOVED_PERMANENTLY","FOUND","MOVED_TEMPORARILY","SEE_OTHER","NOT_MODIFIED","USE_PROXY","TEMPORARY_REDIRECT","PERMANENT_REDIRECT","BAD_REQUEST","UNAUTHORIZED","PAYMENT_REQUIRED","FORBIDDEN","NOT_FOUND","METHOD_NOT_ALLOWED","NOT_ACCEPTABLE","PROXY_AUTHENTICATION_REQUIRED","REQUEST_TIMEOUT","CONFLICT","GONE","LENGTH_REQUIRED","PRECONDITION_FAILED","PAYLOAD_TOO_LARGE","REQUEST_ENTITY_TOO_LARGE","URI_TOO_LONG","REQUEST_URI_TOO_LONG","UNSUPPORTED_MEDIA_TYPE","REQUESTED_RANGE_NOT_SATISFIABLE","EXPECTATION_FAILED","I_AM_A_TEAPOT","INSUFFICIENT_SPACE_ON_RESOURCE","METHOD_FAILURE","DESTINATION_LOCKED","UNPROCESSABLE_ENTITY","LOCKED","FAILED_DEPENDENCY","TOO_EARLY","UPGRADE_REQUIRED","PRECONDITION_REQUIRED","TOO_MANY_REQUESTS","REQUEST_HEADER_FIELDS_TOO_LARGE","UNAVAILABLE_FOR_LEGAL_REASONS","INTERNAL_SERVER_ERROR","NOT_IMPLEMENTED","BAD_GATEWAY","SERVICE_UNAVAILABLE","GATEWAY_TIMEOUT","HTTP_VERSION_NOT_SUPPORTED","VARIANT_ALSO_NEGOTIATES","INSUFFICIENT_STORAGE","LOOP_DETECTED","BANDWIDTH_LIMIT_EXCEEDED","NOT_EXTENDED","NETWORK_AUTHENTICATION_REQUIRED"]}}}}}}}},"components":{"schemas":{"TelemetryDto":{"type":"object","properties":{"ip":{"type":"string"},"hostname":{"type":"string"},"cpuUsage":{"type":"number","format":"double"},"hostDiskFree":{"type":"number","format":"double"},"microphoneState":{"type":"string"},"screenCaptureAllowed":{"type":"boolean"},"audioCaptureAllowed":{"type":"boolean"},"dataDate":{"type":"string","format":"date-time"}}},"DeviceDto":{"type":"object","properties":{"hostName":{"type":"string"},"type":{"type":"string","enum":["Laptop","Desktop","Tablet","Smatphone"]},"os":{"type":"string"},"macAddress":{"type":"string"}}},"TelemetryInfoDto":{"type":"object","properties":{"ip":{"type":"string"},"hostname":{"type":"string"},"hostDiskFree":{"type":"number","format":"double"},"microphoneState":{"type":"string"},"screenCaptureAllowed":{"type":"boolean"},"audioCaptureAllowed":{"type":"boolean"},"dataDate":{"type":"string","format":"date-time"}}},"DeviceInfoDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"hostName":{"type":"string"},"type":{"type":"string","enum":["Laptop","Desktop","Tablet","Smatphone"]},"os":{"type":"string"},"macAddress":{"type":"string"}}}}}}