# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/DeviceDto.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import ar.edu.utn.frc.tup.lciii.model.DeviceType;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder(toBuilder = true)
public class DeviceDto {
    private String hostName;
    private DeviceType type;
    private String os;
    private String macAddress;
}